online_inference/requirements.txt
uvicorn==0.13.4
numpy==1.18.5
pandas==1.0.5
pydantic==1.8.1
pyyaml==5.3.1
scikit-learn==0.23.2
fastapi==0.64.0